Allow InterruptCtrlFiber to set interrupt triggered type when add interrupt
